agentic-workflow/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-workflow-agent"
version = "0.1.0"
description = "Standalone Elastic-themed AI agent project for agentic workflow demos."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "lijiarui" },
]
dependencies = [
  "elasticsearch>=8.0,<9",
  "httpx>=0.27,<1",
  "openai>=1.0,<2",
  "pydantic>=2.8,<3",
  "PyYAML>=6.0,<7",
  "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11,<2",
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.24,<1",
  "pytest-cov>=5,<6",
  "pytest-mock>=3.14,<4",
  "ruff>=0.6.9,<1",
  "types-PyYAML>=6.0.12,<7",
]

[project.scripts]
elastic-agent = "agentic_workflow_agent.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["agentic_workflow_agent*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config --cov=agentic_workflow_agent --cov-branch"
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["agentic_workflow_agent"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agentic_workflow_agent"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
files = ["src", "tests"]
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true
